vppinfra: improve bytes display in memory tracing
* Modify n_bytes from u32 to u64 to accurately track allocation
greater than ~4GB

* Update the format function so that the n_bytes values is displayed
as an unsigned long long integer (as opposed to a signed integer),
and displays a max of 10 digits rather than 9 digits.

Type: improvement

+++ src/vppinfra/mem.h
@@ -228,7 +228,7 @@ typedef struct
   u32 n_allocations;
 
   /* Count of bytes allocated with this traceback. */
-  u32 n_bytes;
+  u64 n_bytes;
 
   /* Offset of this item */
   uword offset;
